+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - **patch-list:** show show patch chek-out state as expected. Previously a checked out patch would not have the associated checkmark denoting its state shown in the patches list unless a check out AND a list refresh was shown.
 - **commands:** don't fail checking out patch branch if the branch already existed but was referring to a different revision than the one we're attempting to check out
+- **webview:** make webview restoration across sessions more robust
 - **config:** watch _user-defined_ path to Radicle CLI binary for changes too. Previously only the default paths per OS were being watched.
 - **onboarding:** detect Radicle CLI binary installation change even if file or parent directory tree is missing on extension's initialization